refactoring: AuthenticationConverter 추상화
