+++ src/main/java/nextstep/subway/auth/ui/AuthenticationConverter.java
@@ -1,0 +1,11 @@
+package nextstep.subway.auth.ui;
+
+import nextstep.subway.auth.domain.AuthenticationToken;
+
+import javax.servlet.http.HttpServletRequest;
+import java.io.IOException;
+
+public interface AuthenticationConverter {
+
+    AuthenticationToken convert(HttpServletRequest request) throws IOException;
+}
